# contracts/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Jobseeker Copilot - Authentication Service API",
    "description" : "Authentication and user management microservice providing JWT-based stateless authentication, user registration, and token validation.",
    "version" : "2.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "tags" : [ {
    "name" : "Authentication",
    "description" : "User registration, login, and token management"
  }, {
    "name" : "Users",
    "description" : "User lookup and management operations"
  } ],
  "paths" : {
    "/api/auth/register" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Register a new user",
        "description" : "Registers a new user account with the provided details. Returns a success message on completion.",
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Rotate a refresh token",
        "description" : "Consumes a one-time refresh token and returns a new access/refresh token pair.",
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/password-reset/request" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Request a password-reset email",
        "description" : "Returns the same accepted response whether or not an active account exists.",
        "operationId" : "requestPasswordReset",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/password-reset/complete" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Complete a password reset",
        "description" : "Atomically consumes a password-reset token, changes the password, and revokes all sessions.",
        "operationId" : "completePasswordReset",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordResetCompletionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Revoke the current session",
        "description" : "Immediately revokes the session identified by the bearer access token.",
        "operationId" : "logout",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/auth/login" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Authenticate user and get JWT token",
        "description" : "Authenticates a user with username and password credentials. Returns a JWT token and user details upon successful authentication.",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/registration-requirements" : {
      "get" : {
        "tags" : [ "Authentication" ],
        "summary" : "Get the current registration legal requirements",
        "description" : "Returns the server-authoritative legal version, minimum age and reviewed HTTPS document URLs required for registration.",
        "operationId" : "getRegistrationLegalRequirements",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegistrationLegalRequirements"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me" : {
      "get" : {
        "tags" : [ "Authentication" ],
        "summary" : "Get current authenticated user",
        "description" : "Returns the current user account. Both a trusted service identity and the user's bearer JWT are required.",
        "operationId" : "getCurrentUser",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserAccountResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ],
          "serviceToken" : [ ]
        } ]
      }
    },
    "/api/auth/account/export" : {
      "get" : {
        "tags" : [ "Account lifecycle" ],
        "summary" : "Export personal account data",
        "description" : "Requires a session created by sign-in within the last 15 minutes. The synchronous JSON response is not retained by the service.",
        "operationId" : "exportPersonalData",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Complete no-store personal-data export",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PersonalDataExport"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ],
          "serviceToken" : [ ]
        } ]
      }
    },
    "/.well-known/jwks.json" : {
      "get" : {
        "tags" : [ "Token verification" ],
        "summary" : "Get public access-token verification keys",
        "description" : "Returns active and rotation-overlap RSA public keys in JWKS format.",
        "operationId" : "keys",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "array",
                    "items" : {
                      "type" : "object",
                      "additionalProperties" : {
                        "type" : "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/account" : {
      "delete" : {
        "tags" : [ "Account lifecycle" ],
        "summary" : "Delete the current account and owned data",
        "description" : "Disables credentials first and starts an idempotent, retryable cross-service deletion operation. Document content follows DOC-09 recoverable deletion and legal-hold rules.",
        "operationId" : "deleteAccount",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 16,
            "pattern" : "[A-Za-z0-9][A-Za-z0-9._:-]{15,127}"
          }
        } ],
        "responses" : {
          "202" : {
            "description" : "Deletion accepted or safely resumed",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountDeletionResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ],
          "serviceToken" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "RegisterRequest" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          },
          "termsAccepted" : {
            "type" : "boolean",
            "description" : "Must be true after the user has actively accepted the current Terms of Use."
          },
          "privacyNoticeAcknowledged" : {
            "type" : "boolean",
            "description" : "Must be true after the user has acknowledged the current Privacy Notice."
          },
          "ageEligibilityConfirmed" : {
            "type" : "boolean",
            "description" : "Must be true to confirm the registrant is at least 18 years old."
          },
          "legalVersion" : {
            "type" : "string",
            "description" : "Exact reviewed legal-document version displayed to the user.",
            "maxLength" : 64,
            "minLength" : 1
          }
        },
        "required" : [ "ageEligibilityConfirmed", "legalVersion", "privacyNoticeAcknowledged", "termsAccepted" ]
      },
      "RefreshRequest" : {
        "type" : "object",
        "properties" : {
          "refreshToken" : {
            "type" : "string"
          }
        }
      },
      "LoginResponse" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "tokenType" : {
            "type" : "string"
          },
          "expiresIn" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "PasswordResetRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          }
        }
      },
      "PasswordResetCompletionRequest" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string"
          },
          "newPassword" : {
            "type" : "string"
          }
        }
      },
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        }
      },
      "RegistrationLegalRequirements" : {
        "type" : "object",
        "properties" : {
          "legalVersion" : {
            "type" : "string"
          },
          "minimumAge" : {
            "type" : "integer",
            "format" : "int32"
          },
          "termsUrl" : {
            "type" : "string"
          },
          "privacyNoticeUrl" : {
            "type" : "string"
          }
        }
      },
      "UserAccountResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        }
      },
      "JsonNode" : {
        "type" : "object",
        "properties" : {
          "array" : {
            "type" : "boolean"
          },
          "empty" : {
            "type" : "boolean"
          },
          "null" : {
            "type" : "boolean"
          },
          "float" : {
            "type" : "boolean"
          },
          "string" : {
            "type" : "boolean"
          },
          "container" : {
            "type" : "boolean"
          },
          "number" : {
            "type" : "boolean"
          },
          "missingNode" : {
            "type" : "boolean"
          },
          "valueNode" : {
            "type" : "boolean"
          },
          "object" : {
            "type" : "boolean"
          },
          "nodeType" : {
            "type" : "string",
            "enum" : [ "ARRAY", "BINARY", "BOOLEAN", "MISSING", "NULL", "NUMBER", "OBJECT", "POJO", "STRING" ]
          },
          "pojo" : {
            "type" : "boolean"
          },
          "integralNumber" : {
            "type" : "boolean"
          },
          "floatingPointNumber" : {
            "type" : "boolean"
          },
          "short" : {
            "type" : "boolean"
          },
          "int" : {
            "type" : "boolean"
          },
          "long" : {
            "type" : "boolean"
          },
          "double" : {
            "type" : "boolean"
          },
          "bigDecimal" : {
            "type" : "boolean"
          },
          "bigInteger" : {
            "type" : "boolean"
          },
          "textual" : {
            "type" : "boolean",
            "deprecated" : true
          },
          "boolean" : {
            "type" : "boolean"
          },
          "binary" : {
            "type" : "boolean"
          },
          "embeddedValue" : {
            "type" : "boolean"
          }
        }
      },
      "PersonalDataExport" : {
        "type" : "object",
        "properties" : {
          "schemaVersion" : {
            "type" : "string"
          },
          "generatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "account" : {
            "$ref" : "#/components/schemas/UserAccountResponse"
          },
          "registrationLegalAcceptance" : {
            "$ref" : "#/components/schemas/RegistrationLegalAcceptanceResponse"
          },
          "profile" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "applications" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "documents" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "payments" : {
            "$ref" : "#/components/schemas/JsonNode"
          }
        }
      },
      "RegistrationLegalAcceptanceResponse" : {
        "type" : "object",
        "properties" : {
          "legalVersion" : {
            "type" : "string"
          },
          "termsAccepted" : {
            "type" : "boolean"
          },
          "privacyNoticeAcknowledged" : {
            "type" : "boolean"
          },
          "ageEligibilityConfirmed" : {
            "type" : "boolean"
          },
          "acceptedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "AccountDeletionResponse" : {
        "type" : "object",
        "properties" : {
          "operationId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PENDING", "RETRY_REQUIRED", "COMPLETED" ]
          },
          "acceptedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "completedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "serviceToken" : {
        "type" : "apiKey",
        "name" : "X-Service-Token",
        "in" : "header"
      }
    }
  }
}