# src/main/java/com/jobseekercopilot/authenticationservice/model/AccountDeletionResponse.java
package com.jobseekercopilot.authenticationservice.model;

import java.time.Instant;

public record AccountDeletionResponse(
        String operationId,
        AccountDeletionStatus status,
        Instant acceptedAt,
        Instant completedAt) {
}